% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_list.R
\name{cal_cov}
\alias{cal_cov}
\title{Calculate standard error (SE)}
\usage{
cal_cov(y_data, theta_mat, directed, network_stats, node_attr = NA)
}
\arguments{
\item{y_data}{The sequence of dynamic networks}

\item{network_stats}{The network statistics for both formation and dissolution models. See search.ergmTerms() for a comprehensive list from the library(ergm).}
}
\value{
Returns a list of results
}
\description{
This function calculates the standard error
}
\examples{
library(CPDstergm)

}
