% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_list.R
\name{CPD_STERGM}
\alias{CPD_STERGM}
\title{Change point detection for a sequence of dynamic networks}
\usage{
CPD_STERGM(
  y_data,
  directed,
  network_stats,
  node_attr = NA,
  list_of_lambda = 10^c(-2:7),
  ADMM_iteration = 200,
  theta_iter = 20,
  z_iter = 20,
  theta_tol = 0.001,
  ADMM_tol = 1e-07,
  threshold_alpha = 0.1,
  verbose = FALSE
)
}
\arguments{
\item{y_data}{The sequence of dynamic networks}

\item{directed}{Whether the networks are directed or not.}

\item{network_stats}{The network statistics for both formation and dissolution models. See search.ergmTerms() for a comprehensive list from the library(ergm).}

\item{node_attr}{The nodal attributes.}

\item{list_of_lambda}{The list of tuning parameters \eqn{\lambda}.}

\item{ADMM_iteration}{The learning iteration for ADMM.}

\item{theta_iter}{The learning iteration for \eqn{\bm{\theta}}.}

\item{z_iter}{The learning iteration for \eqn{\bm{z}}.}

\item{theta_tol}{The tolerance for \eqn{\bm{\theta}} stopping criteria.}

\item{ADMM_tol}{The tolerance for ADMM stopping criteria.}

\item{threshold_alpha}{The alpha level for the data-driven threshold to declare the change points}

\item{verbose}{If it is TRUE, the information at each ADMM iteration is printed.}
}
\value{
Returns a list of results including (1) estimated change points, (2) BIC value, (3) sequential parameter change after standardization or \eqn{\Delta \hat{\bm{\zeta}}}, (4) data-driven threshold to declare the change points, and (5) the learned parameters \eqn{\bm{\theta}}.
}
\description{
This function detects the change points for a sequence of dynamic networks
}
\examples{
library(ecp)
data(DJIA)
market <- DJIA$market
date_vec <- DJIA$dates[1:1138]
rownames(market) <- date_vec

start <- which(date_vec == '2007-01-01')
end <- which(date_vec == '2010-01-04')
date_range <- start:end
mydata <- array(NA,c(29, 29, length(date_range)))
df <- list()

for(i in 1:length(date_range)){
  temp <- market[date_range[i]:(date_range[i]+3),]
  temp <- ifelse(cor(temp)< 0, 1, 0)
  diag(temp) <- 0
  df[[i]] <- temp
}

result <- CPD_STERGM(df, directed=FALSE, network_stats=c("edges", "triangles"),
                     list_of_lambda=10^c(2:5))

seq_date <- rownames(market[start:end,])
seq_date[result[[1]]]
}
