% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_list.R
\name{CPD_STERGM_batch}
\alias{CPD_STERGM_batch}
\title{Perform change point detection}
\usage{
CPD_STERGM_batch(
  data_list,
  network_stats,
  node_attr = NA,
  list_of_lambda = 10^c(-2:7),
  ADMM_iteration = 200,
  theta_iter = 20,
  z_iter = 20,
  theta_tol = 0.001,
  ADMM_tol = 1e-07,
  true_CP = c(26, 51, 76)
)
}
\arguments{
\item{data_list}{list of networks}

\item{network_stats}{network statistics}

\item{node_attr}{nodal attribute}

\item{list_of_lambda}{list of tunning parameters}

\item{ADMM_iteration}{ADMM learning iteration}

\item{theta_iter}{theta learning iteration}

\item{z_iter}{z learning iteration}

\item{theta_tol}{tolerance for theta stopping criteria}

\item{ADMM_tol}{tolerance for ADMM stopping criteria}

\item{true_CP}{true change points}
}
\value{
result of metrics
}
\description{
Perform change point detection
}
\examples{
# network_stats=c("edges", "mutual")
# sim1_result1 <- CPD_STERGM_batch(SBM_list, network_stats)
# colMeans(sim1_result1)
}
