% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_list.R
\name{Evaluation_ADMM}
\alias{Evaluation_ADMM}
\title{Evaluate the algorithm performance}
\usage{
Evaluation_ADMM(
  H_pos_list,
  H_neg_list,
  y_pos_list,
  y_neg_list,
  theta_mat,
  tau,
  p1,
  p2,
  p,
  n,
  true_CP
)
}
\arguments{
\item{H_pos_list}{change statistics in the formation model}

\item{H_neg_list}{change statistics in the dissolution model}

\item{y_pos_list}{formation network}

\item{y_neg_list}{dissolution network}

\item{theta_mat}{theta parameter}

\item{tau}{number of row}

\item{p1}{number of network statistics in the formation model}

\item{p2}{number of network statistics in the dissolution model}

\item{p}{total number of network statistics}

\item{n}{number of node}

\item{true_CP}{true change points}
}
\value{
result of metrics
}
\description{
Evaluate the algorithm performance
}
