% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_list.R
\name{save_H_y_list}
\alias{save_H_y_list}
\title{Construct input data to the algorithm}
\usage{
save_H_y_list(y_data, network_stats, node_attr = NA)
}
\arguments{
\item{y_data}{network data}

\item{network_stats}{network statistics}

\item{node_attr}{nodal attribute}
}
\value{
list of input data
}
\description{
Construct input data to the algorithm
}
