% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_list.R
\name{sim_SBM_batch}
\alias{sim_SBM_batch}
\title{Simulate networks from SBM}
\usage{
sim_SBM_batch(num_nets = 10, n = 50, rho = 0)
}
\arguments{
\item{num_nets}{number of networks}

\item{n}{number of node}

\item{rho}{correlation coefficient}
}
\value{
list of networks
}
\description{
Simulate networks from SBM
}
\examples{
# set.seed(1)
# SBM_list <- sim_SBM_batch(num_nets=10, n=50, rho=0)
}
