% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_list.R
\name{sim_SBM_list}
\alias{sim_SBM_list}
\title{Simulation of dynamic networks from the Stochastic Block Model}
\usage{
sim_SBM_list(num_nets = 10, n = 50, rho = 0)
}
\arguments{
\item{num_nets}{The number of sequences of dynamic networks.}

\item{n}{The number of node.}

\item{rho}{The correlation coefficient between consecutive time points for a dyad.}
}
\value{
Returns a list of multiple sequences of dynamic networks.
}
\description{
This function simulates multiple sequences of dynamic networks from the Stochastic Block Model (SBM).
}
\examples{
set.seed(1)
SBM_list <- sim_SBM_list(num_nets=1, n=50, rho=0)
}
