% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_list.R
\name{sim_STERGM_batch}
\alias{sim_STERGM_batch}
\title{Simulate networks from STERGM}
\usage{
sim_STERGM_batch(
  num_nets = 10,
  n = 50,
  network_stats,
  coefs_pos,
  coefs_neg,
  y1_stats,
  node_attr = NA
)
}
\arguments{
\item{num_nets}{number of networks}

\item{n}{number of node}

\item{network_stats}{network statistics}

\item{coefs_pos}{coefficients for the formation model}

\item{coefs_neg}{coefficients for the dissolution model}

\item{y1_stats}{edges for the first network}

\item{node_attr}{nodal attribute}
}
\value{
list of networks
}
\description{
Simulate networks from STERGM
}
\examples{
# coefs_pos <- matrix(c(-1, -1, -1, -1, -2,  1, -2,  1), nrow=2, ncol=4, byrow = T)
# coefs_neg <- matrix(c( -1,  -1, -1,  -1, -2,  -1, -2,  -1), nrow=2, ncol=4, byrow = T)

# y1_stats <- 250
# network_stats=c("edges", "mutual")

# set.seed(1)
# STERGM_list <- sim_STERGM_batch(num_nets=10, n=50, network_stats,
#                                 coefs_pos, coefs_neg, y1_stats, node_attr=NA)
}
