% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_list.R
\name{sim_STERGM_list}
\alias{sim_STERGM_list}
\title{Simulation of dynamic networks from the Separable Temporal Exponential-family Random Graph Model}
\usage{
sim_STERGM_list(
  num_nets = 1,
  n = 50,
  network_stats,
  coefs_pos,
  coefs_neg,
  y1_stats,
  node_attr = NA
)
}
\arguments{
\item{num_nets}{The number of sequences of dynamic networks.}

\item{n}{The number of node.}

\item{network_stats}{The network statistics for both formation and dissolution models. See search.ergmTerms() for a comprehensive list from the library(ergm).}

\item{coefs_pos}{The coefficients for the formation model.}

\item{coefs_neg}{The coefficients for the dissolution model.}

\item{y1_stats}{The number of edges for the first network.}

\item{node_attr}{The nodal attributes.}
}
\value{
Returns a list of multiple sequences of dynamic networks
}
\description{
This function simulates multiple sequences of dynamic networks, from the Separable Temporal Exponential-family Random Graph Model (STERGM).
}
\examples{
coefs_pos <- matrix(c(-1, -1, -1, -1,
                      -2,  1, -2,  1), nrow=2, ncol=4, byrow=TRUE)
coefs_neg <- matrix(c(-1, -1, -1, -1,
                      -2, -1, -2, -1), nrow=2, ncol=4, byrow=TRUE)

set.seed(1)
STERGM_list <- sim_STERGM_list(num_nets=1, n=50, network_stats=c("edges", "mutual"),
                               coefs_pos, coefs_neg, y1_stats=250, node_attr=NA)
}
