% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_list.R
\name{CPD_STERGM_list}
\alias{CPD_STERGM_list}
\title{Change point detection and performance evaluation for multiple sequences of dynamic networks}
\usage{
CPD_STERGM_list(
  data_list,
  directed,
  network_stats,
  node_attr = NA,
  list_of_lambda = 10^c(-2:7),
  ADMM_iteration = 200,
  theta_iter = 20,
  z_iter = 20,
  theta_tol = 0.001,
  ADMM_tol = 1e-07,
  true_CP = c(26, 51, 76),
  verbose = FALSE
)
}
\arguments{
\item{data_list}{The list of multiple sequences of dynamic networks}

\item{directed}{Whether the networks are directed or not.}

\item{network_stats}{The network statistics for both formation and dissolution models. See search.ergmTerms() for a comprehensive list from the library(ergm).}

\item{node_attr}{The nodal attributes.}

\item{list_of_lambda}{The list of tuning parameters \eqn{\lambda}.}

\item{ADMM_iteration}{The learning iteration for ADMM.}

\item{theta_iter}{The learning iteration for \eqn{\bm{\theta}}.}

\item{z_iter}{The learning iteration for \eqn{\bm{z}}.}

\item{theta_tol}{The tolerance for \eqn{\bm{\theta}} stopping criteria.}

\item{ADMM_tol}{The tolerance for ADMM stopping criteria.}

\item{true_CP}{The true change points.}

\item{verbose}{If it is TRUE, the information at each ADMM iteration is printed.}
}
\value{
Returns the four evaluation metrics for each sequence of dynamic networks. The four evaluation metrics are (1) absolute error \eqn{|\hat{K}-K|}, (2) one-sided Hausdorff distance \eqn{d(\hat{\mathcal{C}}|\mathcal{C})}, (3) one-sided Hausdorff distance \eqn{d(\mathcal{C}|\hat{\mathcal{C}})}, and (4) coverage metric \eqn{C(\mathcal{G},\mathcal{G'})}.
}
\description{
This function detects change points for multiple sequences of dynamic networks and perform model selection based on the BIC values.
}
\examples{
set.seed(1)
SBM_list <- sim_SBM_list(num_nets=1, n=50, rho=0)
network_stats=c("edges", "mutual")
sim1_result1 <- CPD_STERGM_list(SBM_list, directed=TRUE, network_stats, list_of_lambda=10^c(3:7))
}
