% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MITphone}
\alias{MITphone}
\title{The MIT Cellphone Data}
\format{
A list with 232 matrices of dimension of 96 by 96.
}
\usage{
MITphone
}
\description{
The Massachusetts Institute of Technology (MIT) cellphone data consists of the human interactions via cellphone activity, among 96 participants for a duration of 232 days. The data were taken from 15-Sept-2004 to 4-May-2005 inclusive, which covers the winter and spring vacations in the MIT 2004-2005 academic calendar.
}
\examples{
data("MITphone")
}
\references{
Eagle, Nathan. "Reality mining: sensing complex social systems." Personal and ubiquitous computing 10.4 (2006): 255-268.
}
\keyword{datasets}
