% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_list.R
\name{save_H_y_list}
\alias{save_H_y_list}
\title{Generation of input data from dynamic networks}
\usage{
save_H_y_list(y_data, directed, network_stats, node_attr = NA)
}
\arguments{
\item{y_data}{A sequence of dynamic networks.}

\item{directed}{Whether the networks are directed or not.}

\item{network_stats}{The network statistics for both formation and dissolution models. See search.ergmTerms() for a comprehensive list from the library(ergm).}

\item{node_attr}{The nodal attributes.}
}
\value{
Returns a list of change statistics and dynamic networks for both formation and dissolution models.
}
\description{
This function generates the change statistics with formation and dissolution networks, and save them as input data.
}
\examples{
set.seed(1)
SBM_list <- sim_SBM_list(num_nets=1, n=50, rho=0)
input_data <- save_H_y_list(SBM_list[[1]], directed=TRUE, network_stats=c("edges", "mutual"))
}
